// types.ts
export interface FrameOptions {
  text: string;
  frameColorStart: string;
  textColor: string;
  font: string;
  fontSize: number;
  textRotation: number;
  imageX: number;
  imageY: number;
  imageScale: number;
  letterSpacing: number;
  filter: string;
  frameThickness: number; 
}

export interface User {
  name: string;
  picture: string;
}